Move reroute logic
